feat: total order for roles
This commit adds the ability to compare `UserRole::User` and
`UserRole::Admin via` <,>,<=,>=,==.

It also adds an almost complete test suite for said functionality.

When creating a new user, you can now also pass in a `role:
"Admin|User"` parameter in the body. If you outrank or are of equal rank
to this role, the new user will take that role.