fix: remove config.toml
We temporarily solved the fedora build issue by separating
fingerprinting from the default feature to its own.

This means frontend development can continue on fedora. It does not
solve the issue at its core - fedora still can not build the
fingerprinting feature, but the default one.
We have to wait for the dependency authors to fix the issue.